+++ src-ui/js/ui/Misc.js
@@ -202,6 +202,7 @@ function toBGimage(pid) {
 		"tontonbeya",
 		"trainstations",
 		"tslither",
+		"turnaround",
 		"voxas",
 		"vslither",
 		"walllogic",